+++ src/Components/ContainerBottom.css
@@ -20,7 +20,7 @@
 	padding-top:1%;
 	padding-bottom:1%;
 
-	/* background-image:url(https://static.tumblr.com/maopbtg/E9Bmgtoht/lines.png), url(https://static.tumblr.com/maopbtg/nBUmgtogx/paper.png); */
+	background-image:url(https://static.tumblr.com/maopbtg/E9Bmgtoht/lines.png), url(https://static.tumblr.com/maopbtg/nBUmgtogx/paper.png);
 
 
   background-repeat:repeat-y, repeat;
